Add playtime ranking function and update leaderboards.py
A function to get playtime rank was added to a new file, dc_utils/localstats.py. Leaderboards.py was updated to include this playtime ranking related to fetching last message from the channel were removed, and a new command to update playtime rank was added. PLAYTIME_CHANNEL_ID was also added to the config.py file.