Add Steam avatar fetching and refactoring
Added functions to fetch small and medium sizes of Steam avatars, improving the functionality of integration with Steam API. Also, refactored 'query.py' to set a thumbnail instead of an image in the embed object. Furthermore, files 'test.py' and 'whitelist.py' have been moved to the 'trending/other' directory for better organization.